requirements.txt
altair
numpy
pandas
pydeck
streamlit
mysql-connector-python
workalendar
openpyxl
